refactor: injected account type
